backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
